refactor to share metrics exporter
