+++ pkg/components/appolly/appolly_test.go
@@ -13,7 +13,7 @@ import (
 	"go.opentelemetry.io/obi/pkg/components/ebpf"
 	"go.opentelemetry.io/obi/pkg/components/exec"
 	"go.opentelemetry.io/obi/pkg/components/pipe/global"
-	"go.opentelemetry.io/obi/pkg/export/otel"
+	"go.opentelemetry.io/obi/pkg/export/otel/otelcfg"
 	"go.opentelemetry.io/obi/pkg/obi"
 )
 
@@ -25,7 +25,7 @@ func TestProcessEventsLoopDoesntBlock(t *testing.T) {
 		},
 		&obi.Config{
 			ChannelBufferLen: 1,
-			Traces: otel.TracesConfig{
+			Traces: otelcfg.TracesConfig{
 				TracesEndpoint: "http://something",
 			},
 		},
