+++ pkg/export/otel/metrics.go
@@ -7,12 +7,7 @@ import (
 	"context"
 	"fmt"
 	"log/slog"
-	"maps"
-	"net/url"
-	"os"
 	"slices"
-	"strings"
-	"time"
 
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetricgrpc"
@@ -33,6 +28,7 @@ import (
 	"go.opentelemetry.io/obi/pkg/export/instrumentations"
 	"go.opentelemetry.io/obi/pkg/export/otel/metric"
 	instrument "go.opentelemetry.io/obi/pkg/export/otel/metric/api/metric"
+	"go.opentelemetry.io/obi/pkg/export/otel/otelcfg"
 	"go.opentelemetry.io/obi/pkg/pipe/msg"
 	"go.opentelemetry.io/obi/pkg/pipe/swarm"
 )
@@ -56,22 +52,8 @@ const (
 	TargetInfo               = "target_info"
 	TracesHostInfo           = "traces_host_info"
 
-	UsualPortGRPC = "4317"
-	UsualPortHTTP = "4318"
-
 	AggregationExplicit    = "explicit_bucket_histogram"
 	AggregationExponential = "base2_exponential_bucket_histogram"
-
-	FeatureNetwork          = "network"
-	FeatureNetworkInterZone = "network_inter_zone"
-	FeatureApplication      = "application"
-	FeatureSpan             = "application_span"
-	FeatureSpanOTel         = "application_span_otel"
-	FeatureSpanSizes        = "application_span_sizes"
-	FeatureGraph            = "application_service_graph"
-	FeatureProcess          = "application_process"
-	FeatureApplicationHost  = "application_host"
-	FeatureEBPF             = "ebpf"
 )
 
 // GrafanaHostIDKey is the same attribute Key as HostIDKey, but used for
@@ -86,161 +68,15 @@ var MetricTypes = []string{
 	"messaging.",
 }
 
-type MetricsConfig struct {
-	Interval time.Duration `yaml:"interval" env:"OTEL_EBPF_METRICS_INTERVAL"`
-	// OTELIntervalMS supports metric intervals as specified by the standard OTEL definition.
-	// OTEL_EBPF_METRICS_INTERVAL takes precedence over it.
-	OTELIntervalMS int `env:"OTEL_METRIC_EXPORT_INTERVAL"`
-
-	CommonEndpoint  string `yaml:"-" env:"OTEL_EXPORTER_OTLP_ENDPOINT"`
-	MetricsEndpoint string `yaml:"endpoint" env:"OTEL_EXPORTER_OTLP_METRICS_ENDPOINT"`
-
-	Protocol        Protocol `yaml:"protocol" env:"OTEL_EXPORTER_OTLP_PROTOCOL"`
-	MetricsProtocol Protocol `yaml:"-" env:"OTEL_EXPORTER_OTLP_METRICS_PROTOCOL"`
-
-	// InsecureSkipVerify is not standard, so we don't follow the same naming convention
-	InsecureSkipVerify bool `yaml:"insecure_skip_verify" env:"OTEL_EBPF_INSECURE_SKIP_VERIFY"`
-
-	Buckets              Buckets `yaml:"buckets"`
-	HistogramAggregation string  `yaml:"histogram_aggregation" env:"OTEL_EXPORTER_OTLP_METRICS_DEFAULT_HISTOGRAM_AGGREGATION"`
-
-	ReportersCacheLen int `yaml:"reporters_cache_len" env:"OTEL_EBPF_METRICS_REPORT_CACHE_LEN"`
-
-	// SDKLogLevel works independently from the global LogLevel because it prints GBs of logs in Debug mode
-	// and the Info messages leak internal details that are not usually valuable for the final user.
-	SDKLogLevel string `yaml:"otel_sdk_log_level" env:"OTEL_EBPF_SDK_LOG_LEVEL"`
-
-	// Features of metrics that are can be exported. Accepted values are "application" and "network".
-	// envDefault is provided to avoid breaking changes
-	Features []string `yaml:"features" env:"OTEL_EBPF_METRICS_FEATURES,expand" envDefault:"${OTEL_EBPF_METRIC_FEATURES}"  envSeparator:","`
-
-	// Allows configuration of which instrumentations should be enabled, e.g. http, grpc, sql...
-	Instrumentations []string `yaml:"instrumentations" env:"OTEL_EBPF_METRICS_INSTRUMENTATIONS" envSeparator:","`
-
-	// TTL is the time since a metric was updated for the last time until it is
-	// removed from the metrics set.
-	TTL time.Duration `yaml:"ttl" env:"OTEL_EBPF_METRICS_TTL"`
-
-	AllowServiceGraphSelfReferences bool `yaml:"allow_service_graph_self_references" env:"OTEL_EBPF_ALLOW_SERVICE_GRAPH_SELF_REFERENCES"`
-
-	// OTLPEndpointProvider allows overriding the OTLP Endpoint. It needs to return an endpoint and
-	// a boolean indicating if the endpoint is common for both traces and metrics
-	OTLPEndpointProvider func() (string, bool) `yaml:"-" env:"-"`
-
-	// InjectHeaders allows injecting custom headers to the HTTP OTLP exporter
-	InjectHeaders func(dst map[string]string) `yaml:"-" env:"-"`
-}
-
-func (m MetricsConfig) MarshalYAML() (any, error) {
-	omit := map[string]struct{}{
-		"endpoint": {},
-	}
-	return omitFieldsForYAML(m, omit), nil
-}
-
-func (m *MetricsConfig) GetProtocol() Protocol {
-	if m.MetricsProtocol != "" {
-		return m.MetricsProtocol
-	}
-	if m.Protocol != "" {
-		return m.Protocol
-	}
-	return m.GuessProtocol()
-}
-
-func (m *MetricsConfig) GetInterval() time.Duration {
-	if m.Interval == 0 {
-		return time.Duration(m.OTELIntervalMS) * time.Millisecond
-	}
-	return m.Interval
-}
-
-func (m *MetricsConfig) GuessProtocol() Protocol {
-	// If no explicit protocol is set, we guess it it from the metrics endpoint port
-	// (assuming it uses a standard port or a development-like form like 14317, 24317, 14318...)
-	ep, _, err := parseMetricsEndpoint(m)
-	if err == nil {
-		if strings.HasSuffix(ep.Port(), UsualPortGRPC) {
-			return ProtocolGRPC
-		} else if strings.HasSuffix(ep.Port(), UsualPortHTTP) {
-			return ProtocolHTTPProtobuf
-		}
-	}
-	// Otherwise we return default protocol according to the latest specification:
-	// https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/protocol/exporter.md?plain=1#L53
-	return ProtocolHTTPProtobuf
-}
-
-func (m *MetricsConfig) OTLPMetricsEndpoint() (string, bool) {
-	if m.OTLPEndpointProvider != nil {
-		return m.OTLPEndpointProvider()
-	}
-	return ResolveOTLPEndpoint(m.MetricsEndpoint, m.CommonEndpoint)
-}
-
-// EndpointEnabled specifies that the OTEL metrics node is enabled if and only if
-// either the OTEL endpoint and OTEL metrics endpoint is defined.
-// If not enabled, this node won't be instantiated
-// Reason to disable linting: it requires to be a value despite it is considered a "heavy struct".
-// This method is invoked only once during startup time so it doesn't have a noticeable performance impact.
-func (m *MetricsConfig) EndpointEnabled() bool {
-	ep, _ := m.OTLPMetricsEndpoint()
-	return ep != ""
-}
-
-func (m *MetricsConfig) AnySpanMetricsEnabled() bool {
-	return m.SpanMetricsEnabled() || m.SpanMetricsSizesEnabled()
-}
-
-func (m *MetricsConfig) SpanMetricsSizesEnabled() bool {
-	return slices.Contains(m.Features, FeatureSpanSizes)
-}
-
-func (m *MetricsConfig) SpanMetricsEnabled() bool {
-	return slices.Contains(m.Features, FeatureSpan) || slices.Contains(m.Features, FeatureSpanOTel)
-}
-
-func (m *MetricsConfig) InvalidSpanMetricsConfig() bool {
-	return slices.Contains(m.Features, FeatureSpan) && slices.Contains(m.Features, FeatureSpanOTel)
-}
-
-func (m *MetricsConfig) HostMetricsEnabled() bool {
-	return slices.Contains(m.Features, FeatureApplicationHost)
-}
-
-func (m *MetricsConfig) ServiceGraphMetricsEnabled() bool {
-	return slices.Contains(m.Features, FeatureGraph)
-}
-
-func (m *MetricsConfig) OTelMetricsEnabled() bool {
-	return slices.Contains(m.Features, FeatureApplication)
-}
-
-func (m *MetricsConfig) NetworkMetricsEnabled() bool {
-	return m.NetworkFlowBytesEnabled() || m.NetworkInterzoneMetricsEnabled()
-}
-
-func (m *MetricsConfig) NetworkFlowBytesEnabled() bool {
-	return slices.Contains(m.Features, FeatureNetwork)
-}
-
-func (m *MetricsConfig) NetworkInterzoneMetricsEnabled() bool {
-	return slices.Contains(m.Features, FeatureNetworkInterZone)
-}
-
-func (m *MetricsConfig) Enabled() bool {
-	return m.EndpointEnabled() && (m.OTelMetricsEnabled() || m.AnySpanMetricsEnabled() || m.NetworkMetricsEnabled())
-}
-
 // MetricsReporter implements the graph node that receives request.Span
 // instances and forwards them as OTEL metrics.
 type MetricsReporter struct {
 	ctx        context.Context
-	cfg        *MetricsConfig
+	cfg        *otelcfg.MetricsConfig
 	hostID     string
 	attributes *attributes.AttrSelector
 	exporter   sdkmetric.Exporter
-	reporters  ReporterPool[*svc.Attrs, *Metrics]
+	reporters  otelcfg.ReporterPool[*svc.Attrs, *Metrics]
 	pidTracker PidServiceTracker
 	is         instrumentations.InstrumentationSelection
 
@@ -301,7 +137,7 @@ type Metrics struct {
 
 func ReportMetrics(
 	ctxInfo *global.ContextInfo,
-	cfg *MetricsConfig,
+	cfg *otelcfg.MetricsConfig,
 	selectorCfg *attributes.SelectorConfig,
 	input *msg.Queue[[]request.Span],
 	processEventCh *msg.Queue[exec.ProcessEvent],
@@ -310,7 +146,7 @@ func ReportMetrics(
 		if !cfg.Enabled() {
 			return swarm.EmptyRunFunc()
 		}
-		SetupInternalOTELSDKLogger(cfg.SDKLogLevel)
+		otelcfg.SetupInternalOTELSDKLogger(cfg.SDKLogLevel)
 
 		mr, err := newMetricsReporter(
 			ctx,
@@ -340,7 +176,7 @@ func ReportMetrics(
 func newMetricsReporter(
 	ctx context.Context,
 	ctxInfo *global.ContextInfo,
-	cfg *MetricsConfig,
+	cfg *otelcfg.MetricsConfig,
 	selectorCfg *attributes.SelectorConfig,
 	input *msg.Queue[[]request.Span],
 	processEventCh *msg.Queue[exec.ProcessEvent],
@@ -411,7 +247,7 @@ func newMetricsReporter(
 			request.SpanOTELGetters, mr.attributes.For(attributes.GPUKernelBlockSize))
 	}
 
-	mr.reporters = NewReporterPool[*svc.Attrs, *Metrics](cfg.ReportersCacheLen, cfg.TTL, timeNow,
+	mr.reporters = otelcfg.NewReporterPool[*svc.Attrs, *Metrics](cfg.ReportersCacheLen, cfg.TTL, timeNow,
 		func(id svc.UID, v *Metrics) {
 			llog := log.With("service", id)
 			llog.Debug("evicting metrics reporter from cache")
@@ -483,7 +319,7 @@ func (mr *MetricsReporter) otelMetricOptions(mlog *slog.Logger) []metric.Option 
 }
 
 func (mr *MetricsReporter) usesLegacySpanNames() bool {
-	return slices.Contains(mr.cfg.Features, FeatureSpan)
+	return slices.Contains(mr.cfg.Features, otelcfg.FeatureSpan)
 }
 
 func (mr *MetricsReporter) spanMetricsLatencyName() string {
@@ -733,7 +569,7 @@ func (mr *MetricsReporter) newMetricsInstance(service *svc.Attrs) Metrics {
 	var resourceAttributes []attribute.KeyValue
 	if service != nil {
 		mlog = mlog.With("service", service)
-		resourceAttributes = append(GetAppResourceAttrs(mr.hostID, service), ResourceAttrsFromEnv(service)...)
+		resourceAttributes = append(otelcfg.GetAppResourceAttrs(mr.hostID, service), otelcfg.ResourceAttrsFromEnv(service)...)
 	}
 	mlog.Debug("creating new Metrics reporter")
 	resources := resource.NewWithAttributes(semconv.SchemaURL, resourceAttributes...)
@@ -804,7 +640,7 @@ func (mr *MetricsReporter) newMetricSet(service *svc.Attrs) (*Metrics, error) {
 	return &m, nil
 }
 
-func isExponentialAggregation(mc *MetricsConfig, mlog *slog.Logger) bool {
+func isExponentialAggregation(mc *otelcfg.MetricsConfig, mlog *slog.Logger) bool {
 	switch mc.HistogramAggregation {
 	case AggregationExponential:
 		return true
@@ -819,29 +655,29 @@ func isExponentialAggregation(mc *MetricsConfig, mlog *slog.Logger) bool {
 }
 
 // TODO: unify in single exporter for most metrics
-func InstantiateMetricsExporter(ctx context.Context, cfg *MetricsConfig, log *slog.Logger) (sdkmetric.Exporter, error) {
+func InstantiateMetricsExporter(ctx context.Context, cfg *otelcfg.MetricsConfig, log *slog.Logger) (sdkmetric.Exporter, error) {
 	var err error
 	var exporter sdkmetric.Exporter
 	switch proto := cfg.GetProtocol(); proto {
-	case ProtocolHTTPJSON, ProtocolHTTPProtobuf, "": // zero value defaults to HTTP for backwards-compatibility
+	case otelcfg.ProtocolHTTPJSON, otelcfg.ProtocolHTTPProtobuf, "": // zero value defaults to HTTP for backwards-compatibility
 		log.Debug("instantiating HTTP MetricsReporter", "protocol", proto)
 		if exporter, err = httpMetricsExporter(ctx, cfg); err != nil {
 			return nil, fmt.Errorf("can't instantiate OTEL HTTP metrics exporter: %w", err)
 		}
-	case ProtocolGRPC:
+	case otelcfg.ProtocolGRPC:
 		log.Debug("instantiating GRPC MetricsReporter", "protocol", proto)
 		if exporter, err = grpcMetricsExporter(ctx, cfg); err != nil {
 			return nil, fmt.Errorf("can't instantiate OTEL GRPC metrics exporter: %w", err)
 		}
 	default:
 		return nil, fmt.Errorf("invalid protocol value: %q. Accepted values are: %s, %s, %s",
-			proto, ProtocolGRPC, ProtocolHTTPJSON, ProtocolHTTPProtobuf)
+			proto, otelcfg.ProtocolGRPC, otelcfg.ProtocolHTTPJSON, otelcfg.ProtocolHTTPProtobuf)
 	}
 	return exporter, nil
 }
 
-func httpMetricsExporter(ctx context.Context, cfg *MetricsConfig) (sdkmetric.Exporter, error) {
-	opts, err := getHTTPMetricEndpointOptions(cfg)
+func httpMetricsExporter(ctx context.Context, cfg *otelcfg.MetricsConfig) (sdkmetric.Exporter, error) {
+	opts, err := otelcfg.HTTPMetricEndpointOptions(cfg)
 	if err != nil {
 		return nil, err
 	}
@@ -852,8 +688,8 @@ func httpMetricsExporter(ctx context.Context, cfg *MetricsConfig) (sdkmetric.Exp
 	return mexp, nil
 }
 
-func grpcMetricsExporter(ctx context.Context, cfg *MetricsConfig) (sdkmetric.Exporter, error) {
-	opts, err := getGRPCMetricEndpointOptions(cfg)
+func grpcMetricsExporter(ctx context.Context, cfg *otelcfg.MetricsConfig) (sdkmetric.Exporter, error) {
+	opts, err := otelcfg.GRPCMetricEndpointOptions(cfg)
 	if err != nil {
 		return nil, err
 	}
@@ -934,7 +770,7 @@ func (mr *MetricsReporter) tracesResourceAttributes(service *svc.Attrs) attribut
 		extraAttrs = append(extraAttrs, k.OTEL().String(v))
 	}
 
-	filteredAttrs := getFilteredMetricResourceAttrs(baseAttrs, mr.userAttribSelection, extraAttrs, MetricTypes)
+	filteredAttrs := otelcfg.GetFilteredAttributesByPrefix(baseAttrs, mr.userAttribSelection, extraAttrs, MetricTypes)
 	return attribute.NewSet(filteredAttrs...)
 }
 
@@ -1170,119 +1006,6 @@ func (mr *MetricsReporter) reportMetrics(_ context.Context) {
 		}
 	}
 	mr.close()
-}
-
-func getHTTPMetricEndpointOptions(cfg *MetricsConfig) (otlpOptions, error) {
-	opts := otlpOptions{Headers: map[string]string{}}
-	log := mlog().With("transport", "http")
-	murl, isCommon, err := parseMetricsEndpoint(cfg)
-	if err != nil {
-		return opts, err
-	}
-	log.Debug("Configuring exporter",
-		"protocol", cfg.Protocol, "metricsProtocol", cfg.MetricsProtocol, "endpoint", murl.Host)
-
-	setMetricsProtocol(cfg)
-	opts.Endpoint = murl.Host
-	if murl.Scheme == "http" || murl.Scheme == "unix" {
-		log.Debug("Specifying insecure connection", "scheme", murl.Scheme)
-		opts.Insecure = true
-	}
-	// If the value is set from the OTEL_EXPORTER_OTLP_ENDPOINT common property, we need to add /v1/metrics to the path
-	// otherwise, we leave the path that is explicitly set by the user
-	opts.URLPath = murl.Path
-	if isCommon {
-		if strings.HasSuffix(opts.URLPath, "/") {
-			opts.URLPath += "v1/metrics"
-		} else {
-			opts.URLPath += "/v1/metrics"
-		}
-	}
-	log.Debug("Specifying path", "path", opts.URLPath)
-
-	if cfg.InsecureSkipVerify {
-		log.Debug("Setting InsecureSkipVerify")
-		opts.SkipTLSVerify = cfg.InsecureSkipVerify
-	}
-
-	if cfg.InjectHeaders != nil {
-		cfg.InjectHeaders(opts.Headers)
-	}
-	maps.Copy(opts.Headers, HeadersFromEnv(envHeaders))
-	maps.Copy(opts.Headers, HeadersFromEnv(envMetricsHeaders))
-
-	return opts, nil
-}
-
-func getGRPCMetricEndpointOptions(cfg *MetricsConfig) (otlpOptions, error) {
-	opts := otlpOptions{Headers: map[string]string{}}
-	log := mlog().With("transport", "grpc")
-	murl, _, err := parseMetricsEndpoint(cfg)
-	if err != nil {
-		return opts, err
-	}
-	log.Debug("Configuring exporter",
-		"protocol", cfg.Protocol, "metricsProtocol", cfg.MetricsProtocol, "endpoint", murl.Host)
-
-	setMetricsProtocol(cfg)
-	opts.Endpoint = murl.Host
-	if murl.Scheme == "http" || murl.Scheme == "unix" {
-		log.Debug("Specifying insecure connection", "scheme", murl.Scheme)
-		opts.Insecure = true
-	}
-	if cfg.InsecureSkipVerify {
-		log.Debug("Setting InsecureSkipVerify")
-		opts.SkipTLSVerify = true
-	}
-
-	if cfg.InjectHeaders != nil {
-		cfg.InjectHeaders(opts.Headers)
-	}
-	maps.Copy(opts.Headers, HeadersFromEnv(envHeaders))
-	maps.Copy(opts.Headers, HeadersFromEnv(envMetricsHeaders))
-
-	return opts, nil
-}
-
-// the HTTP path will be defined from one of the following sources, from highest to lowest priority
-// - the result from any overridden OTLP Provider function
-// - OTEL_EXPORTER_OTLP_METRICS_ENDPOINT, if defined
-// - OTEL_EXPORTER_OTLP_ENDPOINT, if defined
-func parseMetricsEndpoint(cfg *MetricsConfig) (*url.URL, bool, error) {
-	endpoint, isCommon := cfg.OTLPMetricsEndpoint()
-
-	murl, err := url.Parse(endpoint)
-	if err != nil {
-		return nil, isCommon, fmt.Errorf("parsing endpoint URL %s: %w", endpoint, err)
-	}
-	if murl.Scheme == "" || murl.Host == "" {
-		return nil, isCommon, fmt.Errorf("URL %q must have a scheme and a host", endpoint)
-	}
-	return murl, isCommon, nil
-}
-
-// HACK: at the time of writing this, the otelpmetrichttp API does not support explicitly
-// setting the protocol. They should be properly set via environment variables, but
-// if the user supplied the value via configuration file (and not via env vars), we override the environment.
-// To be as least intrusive as possible, we will change the variables if strictly needed
-// TODO: remove this once otelpmetrichttp.WithProtocol is supported
-func setMetricsProtocol(cfg *MetricsConfig) {
-	if _, ok := os.LookupEnv(envMetricsProtocol); ok {
-		return
-	}
-	if _, ok := os.LookupEnv(envProtocol); ok {
-		return
-	}
-	if cfg.MetricsProtocol != "" {
-		os.Setenv(envMetricsProtocol, string(cfg.MetricsProtocol))
-		return
-	}
-	if cfg.Protocol != "" {
-		os.Setenv(envProtocol, string(cfg.Protocol))
-		return
-	}
-	// unset. Guessing it
-	os.Setenv(envMetricsProtocol, string(cfg.GuessProtocol()))
 }
 
 func cleanupMetrics(ctx context.Context, m *Expirer[*request.Span, instrument.Float64Histogram, float64]) {
