+++ pkg/export/otel/metrics_net_test.go
@@ -7,6 +7,8 @@ import (
 	"testing"
 	"time"
 
+	"go.opentelemetry.io/obi/pkg/export/otel/otelcfg"
+
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/otel/attribute"
@@ -19,7 +21,7 @@ import (
 )
 
 func TestMetricAttributes(t *testing.T) {
-	defer restoreEnvAfterExecution()()
+	defer otelcfg.RestoreEnvAfterExecution()()
 	in := &ebpf.Record{
 		NetFlowRecordT: ebpf.NetFlowRecordT{
 			Id: ebpf.NetFlowId{
@@ -47,12 +49,12 @@ func TestMetricAttributes(t *testing.T) {
 			SelectionCfg: map[attributes.Section]attributes.InclusionLists{
 				attributes.NetworkFlow.Section: {Include: []string{"*"}},
 			},
-		}, Metrics: &MetricsConfig{
+		}, Metrics: &otelcfg.MetricsConfig{
 			MetricsEndpoint:   "http://foo",
 			Interval:          10 * time.Millisecond,
 			ReportersCacheLen: 100,
 			TTL:               5 * time.Minute,
-			Features:          []string{FeatureNetwork, FeatureNetworkInterZone},
+			Features:          []string{otelcfg.FeatureNetwork, otelcfg.FeatureNetworkInterZone},
 		}}, msg.NewQueue[[]*ebpf.Record]())
 	require.NoError(t, err)
 
@@ -77,7 +79,7 @@ func TestMetricAttributes(t *testing.T) {
 }
 
 func TestMetricAttributes_Filter(t *testing.T) {
-	defer restoreEnvAfterExecution()()
+	defer otelcfg.RestoreEnvAfterExecution()()
 	in := &ebpf.Record{
 		NetFlowRecordT: ebpf.NetFlowRecordT{
 			Id: ebpf.NetFlowId{
@@ -109,11 +111,11 @@ func TestMetricAttributes_Filter(t *testing.T) {
 					"k8s.dst.name",
 				}},
 			},
-		}, Metrics: &MetricsConfig{
+		}, Metrics: &otelcfg.MetricsConfig{
 			MetricsEndpoint:   "http://foo",
 			Interval:          10 * time.Millisecond,
 			ReportersCacheLen: 100,
-			Features:          []string{FeatureNetwork, FeatureNetworkInterZone},
+			Features:          []string{otelcfg.FeatureNetwork, otelcfg.FeatureNetworkInterZone},
 		}}, msg.NewQueue[[]*ebpf.Record]())
 	require.NoError(t, err)
 
@@ -139,24 +141,24 @@ func TestMetricAttributes_Filter(t *testing.T) {
 }
 
 func TestNetMetricsConfig_Enabled(t *testing.T) {
-	assert.True(t, NetMetricsConfig{Metrics: &MetricsConfig{
-		Features: []string{FeatureApplication, FeatureNetwork}, CommonEndpoint: "foo",
+	assert.True(t, NetMetricsConfig{Metrics: &otelcfg.MetricsConfig{
+		Features: []string{otelcfg.FeatureApplication, otelcfg.FeatureNetwork}, CommonEndpoint: "foo",
 	}}.Enabled())
-	assert.True(t, NetMetricsConfig{Metrics: &MetricsConfig{
-		Features: []string{FeatureNetwork, FeatureApplication}, MetricsEndpoint: "foo",
+	assert.True(t, NetMetricsConfig{Metrics: &otelcfg.MetricsConfig{
+		Features: []string{otelcfg.FeatureNetwork, otelcfg.FeatureApplication}, MetricsEndpoint: "foo",
 	}}.Enabled())
 }
 
 func TestNetMetricsConfig_Disabled(t *testing.T) {
-	fa := []string{FeatureApplication}
-	fn := []string{FeatureNetwork}
-	assert.False(t, NetMetricsConfig{Metrics: &MetricsConfig{Features: fn}}.Enabled())
-	assert.False(t, NetMetricsConfig{Metrics: &MetricsConfig{Features: fn}}.Enabled())
-	assert.False(t, NetMetricsConfig{Metrics: &MetricsConfig{Features: fn}}.Enabled())
+	fa := []string{otelcfg.FeatureApplication}
+	fn := []string{otelcfg.FeatureNetwork}
+	assert.False(t, NetMetricsConfig{Metrics: &otelcfg.MetricsConfig{Features: fn}}.Enabled())
+	assert.False(t, NetMetricsConfig{Metrics: &otelcfg.MetricsConfig{Features: fn}}.Enabled())
+	assert.False(t, NetMetricsConfig{Metrics: &otelcfg.MetricsConfig{Features: fn}}.Enabled())
 	// network feature is not enabled
-	assert.False(t, NetMetricsConfig{Metrics: &MetricsConfig{CommonEndpoint: "foo"}}.Enabled())
-	assert.False(t, NetMetricsConfig{Metrics: &MetricsConfig{MetricsEndpoint: "foo", Features: fa}}.Enabled())
-	assert.False(t, NetMetricsConfig{Metrics: &MetricsConfig{}}.Enabled())
+	assert.False(t, NetMetricsConfig{Metrics: &otelcfg.MetricsConfig{CommonEndpoint: "foo"}}.Enabled())
+	assert.False(t, NetMetricsConfig{Metrics: &otelcfg.MetricsConfig{MetricsEndpoint: "foo", Features: fa}}.Enabled())
+	assert.False(t, NetMetricsConfig{Metrics: &otelcfg.MetricsConfig{}}.Enabled())
 }
 
 func TestGetFilteredNetworkResourceAttrs(t *testing.T) {
