+++ pkg/export/otel/traces.go
@@ -9,13 +9,11 @@ import (
 	"encoding/json"
 	"fmt"
 	"log/slog"
-	"maps"
 	"math"
-	"net/url"
-	"os"
 	"strconv"
-	"strings"
 	"time"
+
+	"go.opentelemetry.io/obi/pkg/export/otel/otelcfg"
 
 	"go.opentelemetry.io/collector/config/configoptional"
 
@@ -51,67 +49,13 @@ import (
 	"go.opentelemetry.io/obi/pkg/export/instrumentations"
 	"go.opentelemetry.io/obi/pkg/pipe/msg"
 	"go.opentelemetry.io/obi/pkg/pipe/swarm"
-	"go.opentelemetry.io/obi/pkg/services"
 )
-
-func tlog() *slog.Logger {
-	return slog.With("component", "otel.TracesReporter")
-}
 
 const reporterName = "go.opentelemetry.io/obi"
 
 type TraceSpanAndAttributes struct {
 	Span       *request.Span
 	Attributes []attribute.KeyValue
-}
-
-type TracesConfig struct {
-	CommonEndpoint string `yaml:"-" env:"OTEL_EXPORTER_OTLP_ENDPOINT"`
-	TracesEndpoint string `yaml:"endpoint" env:"OTEL_EXPORTER_OTLP_TRACES_ENDPOINT"`
-
-	Protocol       Protocol `yaml:"protocol" env:"OTEL_EXPORTER_OTLP_PROTOCOL"`
-	TracesProtocol Protocol `yaml:"-" env:"OTEL_EXPORTER_OTLP_TRACES_PROTOCOL"`
-
-	// Allows configuration of which instrumentations should be enabled, e.g. http, grpc, sql...
-	Instrumentations []string `yaml:"instrumentations" env:"OTEL_EBPF_TRACES_INSTRUMENTATIONS" envSeparator:","`
-
-	// InsecureSkipVerify is not standard, so we don't follow the same naming convention
-	InsecureSkipVerify bool `yaml:"insecure_skip_verify" env:"OTEL_EBPF_INSECURE_SKIP_VERIFY"`
-
-	SamplerConfig services.SamplerConfig `yaml:"sampler"`
-
-	// Configuration options below this line will remain undocumented at the moment,
-	// but can be useful for performance-tuning of some customers.
-	MaxQueueSize int           `yaml:"max_queue_size" env:"OTEL_EBPF_OTLP_TRACES_MAX_QUEUE_SIZE"`
-	BatchTimeout time.Duration `yaml:"batch_timeout" env:"OTEL_EBPF_OTLP_TRACES_BATCH_TIMEOUT"`
-
-	// Configuration options for BackOffConfig of the traces exporter.
-	// See https://github.com/open-telemetry/opentelemetry-collector/blob/main/config/configretry/backoff.go
-	// BackOffInitialInterval the time to wait after the first failure before retrying.
-	BackOffInitialInterval time.Duration `yaml:"backoff_initial_interval" env:"OTEL_EBPF_BACKOFF_INITIAL_INTERVAL"`
-	// BackOffMaxInterval is the upper bound on backoff interval.
-	BackOffMaxInterval time.Duration `yaml:"backoff_max_interval" env:"OTEL_EBPF_BACKOFF_MAX_INTERVAL"`
-	// BackOffMaxElapsedTime is the maximum amount of time (including retries) spent trying to send a request/batch.
-	BackOffMaxElapsedTime time.Duration `yaml:"backoff_max_elapsed_time" env:"OTEL_EBPF_BACKOFF_MAX_ELAPSED_TIME"`
-	ReportersCacheLen     int           `yaml:"reporters_cache_len" env:"OTEL_EBPF_TRACES_REPORT_CACHE_LEN"`
-
-	// SDKLogLevel works independently from the global LogLevel because it prints GBs of logs in Debug mode
-	// and the Info messages leak internal details that are not usually valuable for the final user.
-	SDKLogLevel string `yaml:"otel_sdk_log_level" env:"OTEL_EBPF_SDK_LOG_LEVEL"`
-
-	// OTLPEndpointProvider allows overriding the OTLP Endpoint. It needs to return an endpoint and
-	// a boolean indicating if the endpoint is common for both traces and metrics
-	OTLPEndpointProvider func() (string, bool) `yaml:"-" env:"-"`
-
-	// InjectHeaders allows injecting custom headers to the HTTP OTLP exporter
-	InjectHeaders func(dst map[string]string) `yaml:"-" env:"-"`
-}
-
-func (m TracesConfig) MarshalYAML() (any, error) {
-	omit := map[string]struct{}{
-		"endpoint": {},
-	}
-	return omitFieldsForYAML(m, omit), nil
 }
 
 type SpanAttr struct {
@@ -122,48 +66,8 @@ type SpanAttr struct {
 	Value     [128]uint8
 }
 
-// Enabled specifies that the OTEL traces node is enabled if and only if
-// either the OTEL endpoint and OTEL traces endpoint is defined.
-// If not enabled, this node won't be instantiated
-func (m *TracesConfig) Enabled() bool {
-	return m.CommonEndpoint != "" || m.TracesEndpoint != ""
-}
-
-func (m *TracesConfig) GetProtocol() Protocol {
-	if m.TracesProtocol != "" {
-		return m.TracesProtocol
-	}
-	if m.Protocol != "" {
-		return m.Protocol
-	}
-	return m.guessProtocol()
-}
-
-func (m *TracesConfig) OTLPTracesEndpoint() (string, bool) {
-	if m.OTLPEndpointProvider != nil {
-		return m.OTLPEndpointProvider()
-	}
-	return ResolveOTLPEndpoint(m.TracesEndpoint, m.CommonEndpoint)
-}
-
-func (m *TracesConfig) guessProtocol() Protocol {
-	// If no explicit protocol is set, we guess it it from the metrics enpdoint port
-	// (assuming it uses a standard port or a development-like form like 14317, 24317, 14318...)
-	ep, _, err := parseTracesEndpoint(m)
-	if err == nil {
-		if strings.HasSuffix(ep.Port(), UsualPortGRPC) {
-			return ProtocolGRPC
-		} else if strings.HasSuffix(ep.Port(), UsualPortHTTP) {
-			return ProtocolHTTPProtobuf
-		}
-	}
-	// Otherwise we return default protocol according to the latest specification:
-	// https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/protocol/exporter.md?plain=1#L53
-	return ProtocolHTTPProtobuf
-}
-
 func makeTracesReceiver(
-	cfg TracesConfig,
+	cfg otelcfg.TracesConfig,
 	spanMetricsEnabled bool,
 	ctxInfo *global.ContextInfo,
 	selectorCfg *attributes.SelectorConfig,
@@ -183,7 +87,7 @@ func makeTracesReceiver(
 // TracesReceiver creates a terminal node that consumes request.Spans and sends OpenTelemetry metrics to the configured consumers.
 func TracesReceiver(
 	ctxInfo *global.ContextInfo,
-	cfg TracesConfig,
+	cfg otelcfg.TracesConfig,
 	spanMetricsEnabled bool,
 	selectorCfg *attributes.SelectorConfig,
 	input *msg.Queue[[]request.Span],
@@ -198,7 +102,7 @@ func TracesReceiver(
 }
 
 type tracesOTELReceiver struct {
-	cfg                TracesConfig
+	cfg                otelcfg.TracesConfig
 	ctxInfo            *global.ContextInfo
 	selectorCfg        *attributes.SelectorConfig
 	is                 instrumentations.InstrumentationSelection
@@ -294,7 +198,7 @@ func (tr *tracesOTELReceiver) processSpans(ctx context.Context, exp exporter.Tra
 				continue
 			}
 
-			envResourceAttrs := ResourceAttrsFromEnv(&sample.Span.Service)
+			envResourceAttrs := otelcfg.ResourceAttrsFromEnv(&sample.Span.Service)
 			traces := generateTracesWithAttributes(tr.attributeCache, &sample.Span.Service, envResourceAttrs, tr.ctxInfo.HostID, spanGroup, tr.ctxInfo.ExtraResourceAttributes)
 			err := exp.ConsumeTraces(ctx, traces)
 			if err != nil {
@@ -340,13 +244,13 @@ func (tr *tracesOTELReceiver) provideLoop(ctx context.Context) {
 }
 
 //nolint:cyclop
-func getTracesExporter(ctx context.Context, cfg TracesConfig) (exporter.Traces, error) {
+func getTracesExporter(ctx context.Context, cfg otelcfg.TracesConfig) (exporter.Traces, error) {
 	switch proto := cfg.GetProtocol(); proto {
-	case ProtocolHTTPJSON, ProtocolHTTPProtobuf, "": // zero value defaults to HTTP for backwards-compatibility
+	case otelcfg.ProtocolHTTPJSON, otelcfg.ProtocolHTTPProtobuf, "": // zero value defaults to HTTP for backwards-compatibility
 		slog.Debug("instantiating HTTP TracesReporter", "protocol", proto)
 		var err error
 
-		opts, err := getHTTPTracesEndpointOptions(&cfg)
+		opts, err := otelcfg.HTTPTracesEndpointOptions(&cfg)
 		if err != nil {
 			slog.Error("can't get HTTP traces endpoint options", "error", err)
 			return nil, err
@@ -393,15 +297,15 @@ func getTracesExporter(ctx context.Context, cfg TracesConfig) (exporter.Traces, 
 			exporterhelper.WithCapabilities(consumer.Capabilities{MutatesData: false}),
 			exporterhelper.WithQueue(config.QueueConfig),
 			exporterhelper.WithRetry(config.RetryConfig))
-	case ProtocolGRPC:
+	case otelcfg.ProtocolGRPC:
 		slog.Debug("instantiating GRPC TracesReporter", "protocol", proto)
 		var err error
-		opts, err := getGRPCTracesEndpointOptions(&cfg)
+		opts, err := otelcfg.GRPCTracesEndpointOptions(&cfg)
 		if err != nil {
 			slog.Error("can't get GRPC traces endpoint options", "error", err)
 			return nil, err
 		}
-		endpoint, _, err := parseTracesEndpoint(&cfg)
+		endpoint, _, err := otelcfg.ParseTracesEndpoint(&cfg)
 		if err != nil {
 			slog.Error("can't parse GRPC traces endpoint", "error", err)
 			return nil, err
@@ -437,7 +341,7 @@ func getTracesExporter(ctx context.Context, cfg TracesConfig) (exporter.Traces, 
 		return factory.CreateTraces(ctx, set, config)
 	default:
 		slog.Error(fmt.Sprintf("invalid protocol value: %q. Accepted values are: %s, %s, %s",
-			proto, ProtocolGRPC, ProtocolHTTPJSON, ProtocolHTTPProtobuf))
+			proto, otelcfg.ProtocolGRPC, otelcfg.ProtocolHTTPJSON, otelcfg.ProtocolHTTPProtobuf))
 		return nil, fmt.Errorf("invalid protocol value: %q", proto)
 	}
 }
@@ -458,7 +362,7 @@ func getTraceSettings(dataTypeMetrics component.Type) exporter.Settings {
 	}
 }
 
-func getRetrySettings(cfg TracesConfig) configretry.BackOffConfig {
+func getRetrySettings(cfg otelcfg.TracesConfig) configretry.BackOffConfig {
 	backOffCfg := configretry.NewDefaultBackOffConfig()
 	if cfg.BackOffInitialInterval > 0 {
 		backOffCfg.InitialInterval = cfg.BackOffInitialInterval
@@ -472,17 +376,19 @@ func getRetrySettings(cfg TracesConfig) configretry.BackOffConfig {
 	return backOffCfg
 }
 
+var emptyUID = svc.UID{}
+
 func traceAppResourceAttrs(cache *expirable2.LRU[svc.UID, []attribute.KeyValue], hostID string, service *svc.Attrs) []attribute.KeyValue {
 	// TODO: remove?
 	if service.UID == emptyUID {
-		return GetAppResourceAttrs(hostID, service)
+		return otelcfg.GetAppResourceAttrs(hostID, service)
 	}
 
 	attrs, ok := cache.Get(service.UID)
 	if ok {
 		return attrs
 	}
-	attrs = GetAppResourceAttrs(hostID, service)
+	attrs = otelcfg.GetAppResourceAttrs(hostID, service)
 	cache.Add(service.UID, attrs)
 
 	return attrs
@@ -828,119 +734,6 @@ func spanStartTime(t request.Timings) time.Time {
 		realStart = t.Start
 	}
 	return realStart
-}
-
-// the HTTP path will be defined from one of the following sources, from highest to lowest priority
-// - OTEL_EXPORTER_OTLP_TRACES_ENDPOINT, if defined
-// - OTEL_EXPORTER_OTLP_ENDPOINT, if defined
-// - https://otlp-gateway-${GRAFANA_CLOUD_ZONE}.grafana.net/otlp, if GRAFANA_CLOUD_ZONE is defined
-// If, by some reason, Grafana changes its OTLP Gateway URL in a distant future, you can still point to the
-// correct URL with the OTLP_EXPORTER_... variables.
-func parseTracesEndpoint(cfg *TracesConfig) (*url.URL, bool, error) {
-	endpoint, isCommon := cfg.OTLPTracesEndpoint()
-
-	murl, err := url.Parse(endpoint)
-	if err != nil {
-		return nil, isCommon, fmt.Errorf("parsing endpoint URL %s: %w", endpoint, err)
-	}
-	if murl.Scheme == "" || murl.Host == "" {
-		return nil, isCommon, fmt.Errorf("URL %q must have a scheme and a host", endpoint)
-	}
-	return murl, isCommon, nil
-}
-
-func getHTTPTracesEndpointOptions(cfg *TracesConfig) (otlpOptions, error) {
-	opts := otlpOptions{Headers: map[string]string{}}
-	log := tlog().With("transport", "http")
-
-	murl, isCommon, err := parseTracesEndpoint(cfg)
-	if err != nil {
-		return opts, err
-	}
-
-	log.Debug("Configuring exporter", "protocol",
-		cfg.Protocol, "tracesProtocol", cfg.TracesProtocol, "endpoint", murl.Host)
-	setTracesProtocol(cfg)
-	opts.Scheme = murl.Scheme
-	opts.Endpoint = murl.Host
-	if murl.Scheme == "http" || murl.Scheme == "unix" {
-		log.Debug("Specifying insecure connection", "scheme", murl.Scheme)
-		opts.Insecure = true
-	}
-	// If the value is set from the OTEL_EXPORTER_OTLP_ENDPOINT common property, we need to add /v1/traces to the path
-	// otherwise, we leave the path that is explicitly set by the user
-	opts.URLPath = strings.TrimSuffix(murl.Path, "/")
-	opts.BaseURLPath = strings.TrimSuffix(opts.URLPath, "/v1/traces")
-	if isCommon {
-		opts.URLPath += "/v1/traces"
-		log.Debug("Specifying path", "path", opts.URLPath)
-	}
-
-	if cfg.InsecureSkipVerify {
-		log.Debug("Setting InsecureSkipVerify")
-		opts.SkipTLSVerify = true
-	}
-
-	if cfg.InjectHeaders != nil {
-		cfg.InjectHeaders(opts.Headers)
-	}
-	maps.Copy(opts.Headers, HeadersFromEnv(envHeaders))
-	maps.Copy(opts.Headers, HeadersFromEnv(envTracesHeaders))
-
-	return opts, nil
-}
-
-func getGRPCTracesEndpointOptions(cfg *TracesConfig) (otlpOptions, error) {
-	opts := otlpOptions{Headers: map[string]string{}}
-	log := tlog().With("transport", "grpc")
-	murl, _, err := parseTracesEndpoint(cfg)
-	if err != nil {
-		return opts, err
-	}
-
-	log.Debug("Configuring exporter", "protocol",
-		cfg.Protocol, "tracesProtocol", cfg.TracesProtocol, "endpoint", murl.Host)
-	opts.Endpoint = murl.Host
-	if murl.Scheme == "http" || murl.Scheme == "unix" {
-		log.Debug("Specifying insecure connection", "scheme", murl.Scheme)
-		opts.Insecure = true
-	}
-
-	if cfg.InsecureSkipVerify {
-		log.Debug("Setting InsecureSkipVerify")
-		opts.SkipTLSVerify = true
-	}
-
-	if cfg.InjectHeaders != nil {
-		cfg.InjectHeaders(opts.Headers)
-	}
-	maps.Copy(opts.Headers, HeadersFromEnv(envHeaders))
-	maps.Copy(opts.Headers, HeadersFromEnv(envTracesHeaders))
-	return opts, nil
-}
-
-// HACK: at the time of writing this, the otelptracehttp API does not support explicitly
-// setting the protocol. They should be properly set via environment variables, but
-// if the user supplied the value via configuration file (and not via env vars), we override the environment.
-// To be as least intrusive as possible, we will change the variables if strictly needed
-// TODO: remove this once otelptracehttp.WithProtocol is supported
-func setTracesProtocol(cfg *TracesConfig) {
-	if _, ok := os.LookupEnv(envTracesProtocol); ok {
-		return
-	}
-	if _, ok := os.LookupEnv(envProtocol); ok {
-		return
-	}
-	if cfg.TracesProtocol != "" {
-		os.Setenv(envTracesProtocol, string(cfg.TracesProtocol))
-		return
-	}
-	if cfg.Protocol != "" {
-		os.Setenv(envProtocol, string(cfg.Protocol))
-		return
-	}
-	// unset. Guessing it
-	os.Setenv(envTracesProtocol, string(cfg.guessProtocol()))
 }
 
 func manualSpanAttributes(span *request.Span) []attribute.KeyValue {
